requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=3.14
coverage>=7.2
gunicorn>=21.2
